Remove `benchmark.py`
